Add PID to status output.
