+++ internal/pkg/agent/configuration/upgrade.go
@@ -12,9 +12,6 @@ const (
 
 	// interval between checks for new (upgraded) Agent returning an error status.
 	defaultStatusCheckInterval = 30 * time.Second
-
-	// interval between checks for new (upgraded) Agent crashing.
-	defaultCrashCheckInterval = 10 * time.Second
 )
 
 // UpgradeConfig is the configuration related to Agent upgrades.
@@ -25,7 +22,6 @@ type UpgradeConfig struct {
 type UpgradeWatcherConfig struct {
 	GracePeriod time.Duration             `yaml:"grace_period" config:"grace_period" json:"grace_period"`
 	ErrorCheck  UpgradeWatcherCheckConfig `yaml:"error_check" config:"error_check" json:"error_check"`
-	CrashCheck  UpgradeWatcherCheckConfig `yaml:"crash_check" config:"crash_check" json:"crash_check"`
 }
 type UpgradeWatcherCheckConfig struct {
 	Interval time.Duration `yaml:"interval" config:"interval" json:"interval"`
@@ -37,9 +33,6 @@ func DefaultUpgradeConfig() *UpgradeConfig {
 			GracePeriod: defaultGracePeriodDuration,
 			ErrorCheck: UpgradeWatcherCheckConfig{
 				Interval: defaultStatusCheckInterval,
-			},
-			CrashCheck: UpgradeWatcherCheckConfig{
-				Interval: defaultCrashCheckInterval,
 			},
 		},
 	}
