+++ pkg/control/v2/client/client.go
@@ -105,6 +105,7 @@ type AgentStateInfo struct {
 	Commit    string `json:"commit" yaml:"commit"`
 	BuildTime string `json:"build_time" yaml:"build_time"`
 	Snapshot  bool   `json:"snapshot" yaml:"snapshot"`
+	PID       int    `json:"pid" yaml:"pid"`
 }
 
 // AgentState is the current state of the Elastic Agent.
@@ -470,6 +471,7 @@ func toState(res *cproto.StateResponse) (*AgentState, error) {
 			Commit:    res.Info.Commit,
 			BuildTime: res.Info.BuildTime,
 			Snapshot:  res.Info.Snapshot,
+			PID:       int(res.Info.Pid),
 		},
 		State:        res.State,
 		Message:      res.Message,
